Moved some of the new files around. Moved the private classes to separate files. Removed the FlagsHelper class and only take what we need. Some refactoring in some of the new classes.
