+++ tests/ImageSharp.Tests/FileTestBase.cs
@@ -6,12 +6,11 @@
 namespace ImageSharp.Tests
 {
     using System.Collections.Generic;
-    using System.IO;
 
     /// <summary>
     /// The test base class for reading and writing to files.
     /// </summary>
-    public abstract class FileTestBase
+    public abstract class FileTestBase : TestBase
     {
         /// <summary>
         /// The collection of image files to test against.
@@ -45,26 +44,5 @@ namespace ImageSharp.Tests
             // TestFile.Create(TestImages.Gif.Cheers),
             // TestFile.Create(TestImages.Gif.Giphy) // Perf: Enable for local testing only
         };
-
-        // TODO: Find a better place for this
-        internal const string TestOutputRoot = "TestOutput/";
-
-        protected string CreateOutputDirectory(string path, params string[] pathParts)
-        {
-            var postFix = "";
-            if (pathParts != null && pathParts.Length > 0)
-            {
-                postFix  = "/" + string.Join("/", pathParts);
-            }
-
-            path = TestOutputRoot + path + postFix;
-
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
-
-            return path;
-        }
     }
 }
