+++ tests/ImageSharp.Tests/TestUtilities/Attributes/WithBlankImageAttribute.cs
